Data/Dataviz/Plotly_Dash/finance/requirements.txt
yfinance==0.2.38
dash>=0.38.0
plotly_express
streamlit
